Split session action IPC handlers
